refactor: custom hooks handlers
